2.team_project/requirements.txt
requests
pandas
python-dotenv
playwright
beautifulsoup4
supabase
google-generativeai
google-api-python-client
